# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PatchInstallInput": {
        "additionalProperties": false,
        "properties": {
          "scope": {
            "default": "security",
            "title": "Scope",
            "type": "string"
          },
          "target_mode": {
            "default": "PC",
            "title": "Target Mode",
            "type": "string"
          },
          "targets": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Targets",
            "type": "array"
          }
        },
        "title": "PatchInstallInput",
        "type": "object"
      },
      "PatchStatusInput": {
        "properties": {
          "last_install": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Install"
          },
          "last_result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Result"
          },
          "last_search": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Search"
          },
          "pending_count": {
            "default": 0,
            "title": "Pending Count",
            "type": "integer"
          },
          "pending_critical": {
            "default": 0,
            "title": "Pending Critical",
            "type": "integer"
          },
          "pending_security": {
            "default": 0,
            "title": "Pending Security",
            "type": "integer"
          },
          "reboot_required": {
            "default": false,
            "title": "Reboot Required",
            "type": "boolean"
          },
          "updates": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PatchUpdateItem"
            },
            "title": "Updates",
            "type": "array"
          }
        },
        "title": "PatchStatusInput",
        "type": "object"
      },
      "PatchUpdateItem": {
        "properties": {
          "categories": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "is_security": {
            "default": false,
            "title": "Is Security",
            "type": "boolean"
          },
          "kb": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kb"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "PatchUpdateItem",
        "type": "object"
      },
...
      "SoftwareInventoryInput": {
        "properties": {
          "items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SoftwareItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "title": "SoftwareInventoryInput",
        "type": "object"
      },
      "SoftwareItem": {
        "properties": {
          "install_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Install Date"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Version"
          }
        },
        "required": [
          "name"
        ],
        "title": "SoftwareItem",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "POps central server API. `/api/v1` is the stable surface for integrations; the same endpoints answer under plain `/api` for the panel and existing agents. Authenticate with an API token (`Authorization: Bearer pops_...`, created by a superadmin in Ayarlar → Güvenlik) or a panel login JWT. Paths marked deprecated still work; use their REST names. Reference: docs/api.md.",
    "title": "POps API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/devices/{pc_name}/software": {
      "get": {
        "operationId": "device_software_api_devices__pc_name__software_get",
        "parameters": [
          {
            "in": "path",
            "name": "pc_name",
            "required": true,
            "schema": {
              "title": "Pc Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Device Software",
        "tags": [
          "inventory"
        ]
      }
    },
...
    "/api/v1/patches": {
      "get": {
        "description": "Cihaz başına Windows Update durumu (hiç bildirmeyen cihazlar da listelenir).",
        "operationId": "list_patch_status_api_patches_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Patch Status",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/patches/install": {
      "post": {
        "description": "Ajanlara bekleyen Windows güncellemelerini kurdurur (security: yalnız güvenlik/kritik, all: hepsi).\nAjan yeniden başlatmaz; gerekiyorsa reboot_required bildirir.",
        "operationId": "install_patches_api_patches_install_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchInstallInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Install Patches",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/patches/scan": {
      "post": {
        "description": "Ajanlardan Windows Update taraması ister (sonuç /api/patches/{pc} ile gelir).",
        "operationId": "scan_patches_api_patches_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchInstallInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Scan Patches",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/patches/{pc_name}": {
      "post": {
        "description": "Cihazın son Windows Update taraması (bekleyen güncellemeler, yeniden başlatma gereksinimi).",
        "operationId": "put_patch_status_api_patches__pc_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "pc_name",
            "required": true,
            "schema": {
              "title": "Pc Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchStatusInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Patch Status",
        "tags": [
          "inventory"
        ]
      }
    },
...
    "/api/v1/software": {
      "get": {
        "description": "Filodaki yazılımlar: ad, yayıncı, sürümler ve kaç cihazda kurulu olduğu.",
        "operationId": "search_software_api_software_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 300,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Software",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/software/devices": {
      "get": {
        "description": "Belirli bir yazılımın kurulu olduğu cihazlar ve sürümleri.",
        "operationId": "software_devices_api_software_devices_get",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Software Devices",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/software/{pc_name}": {
      "post": {
        "description": "Cihazın kurulu yazılım listesinin TAMAMI; önceki liste bununla değiştirilir.",
        "operationId": "put_software_api_software__pc_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "pc_name",
            "required": true,
            "schema": {
              "title": "Pc Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SoftwareInventoryInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Software",
        "tags": [
          "inventory"
        ]
      }
    },
...
  }
}